Agregado de símbolos y ventanas con mensajes

+++ FeriaMaipoGrande/GestionPersonas.xaml.cs
@@ -22,13 +22,16 @@ namespace FeriaMaipoGrande
     /// </summary>
     public partial class GestionPersonas : Window
     {
+        
         public GestionPersonas()
         {
             InitializeComponent();
             listarPersonas();
+            WindowStartupLocation = WindowStartupLocation.CenterScreen;
             
         }
 
+       
         private void btnRegresar_Click(object sender, RoutedEventArgs e)
         {
             MainWindow main = new MainWindow();
@@ -41,15 +44,6 @@ namespace FeriaMaipoGrande
         {
             Persona persona = new Persona();
             dynamic listaPersonas = persona.listarPersonas();
-
-            /*dgListaPersonas.Columns[0].Header = "ID";
-            dgListaPersonas.Columns[0].Header = "Nombre";
-            dgListaPersonas.Columns[2].Header = "Apellido P";
-            dgListaPersonas.Columns[3].Header = "Apellido M";
-            dgListaPersonas.Columns[4].Header = "Dirección";
-            dgListaPersonas.Columns[5].Header = "País";
-            dgListaPersonas.Columns[6].Header = "Ciudad";
-            dgListaPersonas.Columns[7].Header = "Num ID";*/
             dgListaPersonas.ItemsSource = listaPersonas;
 
         }
@@ -58,28 +52,40 @@ namespace FeriaMaipoGrande
 
         private void btnAgregar_Click(object sender, RoutedEventArgs e)
         {
+            // OJO. Hay que modificar las acciones del botón guardar.
+
             if (dgListaPersonas.SelectedIndex == -1){
-                string id, nombre, apellidoP, apellidoM, direccion, ciudad, pais;
-                id = txtID.Text;
-                nombre = txtNombre.Text;
-                apellidoP = txtApellidoP.Text;
-                apellidoM = txtApellidoM.Text;
-                direccion = txtDireccion.Text;
-                ciudad = txtCiudad.Text;
-                pais = txtPais.Text;
-                Persona persona = new Persona();
-                persona.NumIdentificador = id;
-                persona.Nombre = nombre;
-                persona.ApellidoPaterno = apellidoP;
-                persona.ApellidoMaterno = apellidoM;
-                persona.Ciudad = ciudad;
-                persona.Pais = pais;
-                persona.Direccion = direccion;
-                JsonConvert.SerializeObject(persona);
-                persona.crearPersona();
-                MessageBox.Show("Persona agregada correctamente.");
-                LimpiarCampos();
-                listarPersonas();
+                if(!String.IsNullOrEmpty(txtApellidoM.Text) && !String.IsNullOrEmpty(txtDireccion.Text) &&
+                    !String.IsNullOrEmpty(txtApellidoP.Text) && !String.IsNullOrEmpty(txtID.Text) &&
+                    !String.IsNullOrEmpty(txtCiudad.Text) && !String.IsNullOrEmpty(txtNombre.Text) && !String.IsNullOrEmpty(txtPais.Text))
+                {
+                    string id, nombre, apellidoP, apellidoM, direccion, ciudad, pais;
+                    id = txtID.Text;
+                    nombre = txtNombre.Text;
+                    apellidoP = txtApellidoP.Text;
+                    apellidoM = txtApellidoM.Text;
+                    direccion = txtDireccion.Text;
+                    ciudad = txtCiudad.Text;
+                    pais = txtPais.Text;
+                    Persona persona = new Persona();
+                    persona.NumIdentificador = id;
+                    persona.Nombre = nombre;
+                    persona.ApellidoPaterno = apellidoP;
+                    persona.ApellidoMaterno = apellidoM;
+                    persona.Ciudad = ciudad;
+                    persona.Pais = pais;
+                    persona.Direccion = direccion;
+                    JsonConvert.SerializeObject(persona);
+                    persona.crearPersona();
+                    MessageBox.Show("Persona agregada correctamente.");
+                    LimpiarCampos();
+                    listarPersonas();
+                }
+                else
+                {
+                    MessageBox.Show("Debe completar los campos para añadir una persona", "Datos faltantes", MessageBoxButton.OK, MessageBoxImage.Information);
+                }
+                
             }
             else{
                 ModificarPersona();
@@ -92,16 +98,21 @@ namespace FeriaMaipoGrande
         {
             if (dgListaPersonas.SelectedIndex == -1 || dgListaPersonas.SelectedItem.ToString().Equals("{NewItemPlaceholder}"))
             {
-                MessageBox.Show("No hay datos seleccionados, por favor seleccione.");
+                MessageBox.Show("No hay datos seleccionados, por favor seleccione.","Infromación", MessageBoxButton.OK, MessageBoxImage.Information);
             }
             else
             {
-                dynamic persona = JObject.Parse(dgListaPersonas.SelectedItem.ToString());
-                string numID = persona["num_identificador"].ToString();
-                Persona pers = new Persona();
-                pers.eliminarPersona(numID);
-                MessageBox.Show("Persona eliminada correctamente.");
-                listarPersonas();
+                MessageBoxResult resutlado = MessageBox.Show("Está seguro que desea eliminar a esta persona?", "Alerta", MessageBoxButton.YesNo, MessageBoxImage.Warning);
+                if(resutlado == MessageBoxResult.Yes)
+                {
+                    dynamic persona = JObject.Parse(dgListaPersonas.SelectedItem.ToString());
+                    string numID = persona["num_identificador"].ToString();
+                    Persona pers = new Persona();
+                    pers.eliminarPersona(numID);
+                    MessageBox.Show("Persona eliminada correctamente.", "Registro modificado", MessageBoxButton.OK, MessageBoxImage.Information);
+                    listarPersonas();
+                }
+
             }
             listarPersonas();
         }
@@ -110,7 +121,7 @@ namespace FeriaMaipoGrande
         {
             if (dgListaPersonas.SelectedIndex == -1)
             {
-                MessageBox.Show("Seleccione una persona.");
+                MessageBox.Show("Seleccione una persona.", "Información", MessageBoxButton.OK, MessageBoxImage.Exclamation);
             }
             else
             {
